requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
conda==4.8.3
conda-package-handling==1.7.0
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
idna==2.9
IMDbPY==2020.9.25
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lxml==4.6.1
MarkupSafe==1.1.1
matplotlib==3.3.3
menuinst==1.4.16
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
pycosat==0.6.3
pycparser==2.20
Pygments==2.7.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.15.1
pyqt5-plugins==5.15.1.2.0.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.1.3
PyQt5Designer==5.14.1
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
pywin32==227
pywinpty==0.5.7
pyzmq==20.0.0
qt5-applications==5.15.1.2.1
qt5-tools==5.15.1.1.0.1
qtconsole==4.7.7
QtPy==1.9.0
requests==2.23.0
ruamel-yaml==0.15.87
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.46.0
traitlets==5.0.5
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
xgboost==1.3.1
